Moving explicit and implicit micro-experiments from exp03_syr2k to exp04_update_of_c
